requirements.txt
# Core
torch>=2.0
numpy>=1.24
pandas>=2.0
tqdm>=4.65

# CLIP embeddings
open_clip_torch>=2.20

# Colour space conversion
colour-science>=0.4